+++ Components/IRacing/Events/IRacingCompletedLap.cs
@@ -11,6 +11,7 @@ namespace Slipstream.Components.IRacing.Events
         public double SessionTime { get; set; }
         public long CarIdx { get; set; }
         public double LapTime { get; set; }
+        public bool EstimatedLapTime { get; set; }
         public int LapsCompleted { get; set; }
         public float? FuelDelta { get; set; }
         public bool LocalUser { get; set; }
@@ -24,6 +25,7 @@ namespace Slipstream.Components.IRacing.Events
                    SessionTime == lap.SessionTime &&
                    CarIdx == lap.CarIdx &&
                    LapTime == lap.LapTime &&
+                   EstimatedLapTime == lap.EstimatedLapTime &&
                    LapsCompleted == lap.LapsCompleted &&
                    FuelDelta == lap.FuelDelta &&
                    LocalUser == lap.LocalUser &&
@@ -38,6 +40,7 @@ namespace Slipstream.Components.IRacing.Events
             hashCode = hashCode * -1521134295 + SessionTime.GetHashCode();
             hashCode = hashCode * -1521134295 + CarIdx.GetHashCode();
             hashCode = hashCode * -1521134295 + LapTime.GetHashCode();
+            hashCode = hashCode * -1521134295 + EstimatedLapTime.GetHashCode();
             hashCode = hashCode * -1521134295 + LapsCompleted.GetHashCode();
             hashCode = hashCode * -1521134295 + FuelDelta.GetHashCode();
             hashCode = hashCode * -1521134295 + LocalUser.GetHashCode();